Refactor passport conversion logic to separate service
The passport conversion logic has been separated from RASAuthenticationService into a new service, RASPassPortService. The tests have also been updated to accommodate these changes. This improves code organization and makes each service more focused on a specific set of responsibilities.
